feat: cleanup-ui-factory-session-detail-test-split-003 - Preserve dispatch drilldown and durable detail boundary coverage
